Make command reusable
